+++ include/dxc/config.h.cmake
@@ -1,0 +1,2 @@
+/* Disable overriding memory allocators. */
+#cmakedefine DXC_DISABLE_ALLOCATOR_OVERRIDES
